move exports to indts.ts instead of tsx
