+++ packages/x-charts/src/BarChart/index.ts
