+++ packages/x-charts/src/ChartsClipPath/index.ts
