+++ packages/x-charts/src/ChartsSurface/index.ts
@@ -1,0 +1,1 @@
+export * from './ChartsSurface'
