+++ packages/x-charts/src/ResponsiveChartContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './ResponsiveChartContainer';
